Code fixes. View models refactor progress.
