+++ app/src/main/java/dev/ricsarabia/cryptochallenge/ui/main/MainFragment.kt
@@ -13,33 +13,43 @@ import androidx.recyclerview.widget.GridLayoutManager
 import dev.ricsarabia.cryptochallenge.R
 import dev.ricsarabia.cryptochallenge.databinding.MainFragmentBinding
 import dev.ricsarabia.cryptochallenge.domain.Book
-import dev.ricsarabia.cryptochallenge.ui.MainViewModel
+import dev.ricsarabia.cryptochallenge.ui.detail.DetailViewModel
 
 class MainFragment : Fragment() {
-    private lateinit var binding: MainFragmentBinding
-    private val viewModel by lazy { ViewModelProvider(requireActivity()).get(MainViewModel::class.java) }
+    private var _binding: MainFragmentBinding? = null
+    private val binding get() = _binding!!
+
+    private val viewModel by lazy { ViewModelProvider(requireActivity()).get(DetailViewModel::class.java) }
     private val booksAdapter = BooksAdapter { onBookClick(it) }
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        binding = MainFragmentBinding.inflate(inflater, container, false)
+        _binding = MainFragmentBinding.inflate(inflater, container, false)
         return binding.root
     }
 
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        // Init components
+        initViews()
+        initObservers()
+        viewModel.updateBooks()
+    }
+
+    private fun initViews() {
         binding.booksLinearLayout.layoutManager = GridLayoutManager(context, 2)
         binding.booksLinearLayout.adapter = booksAdapter
+    }
 
-        // Init observers
+    private fun initObservers() {
         viewModel.books.observe(viewLifecycleOwner, { booksAdapter.books = it })
         viewModel.loading.observe(viewLifecycleOwner, { binding.mainProgressBar.isVisible = it })
         viewModel.errorMessage.observe(viewLifecycleOwner, { Log.wtf("errorMessage", it) })
-
-        // Retrieving books data
-        viewModel.updateBooks()
     }
 
     private fun onBookClick(book: Book) {
